3.2-crud/stocks_products/requirements.txt
django~=5.1
djangorestframework~=3.15.2
psycopg2-binary

python-dotenv~=1.0.1
django-filter~=24.3